% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult_grad_codes.R
\name{adult_grad_codes}
\alias{adult_grad_codes}
\title{List of federal graduation codes.}
\usage{
adult_grad_codes()
}
\value{
list of federal adult graduation codes
}
\description{
This function returns a list of current adult diploma codes according to federal guidelines.
}
\section{Last Updated By}{

Stevens Dormezil
}

\section{Last Update Date}{

2022/06/09
}

\examples{
adult_grad_codes()

}
\keyword{diploma}
\keyword{federal,}
\keyword{graduation,}
