% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/certificate_codes.R
\name{certificate_codes}
\alias{certificate_codes}
\title{List of certificate of completion codes.}
\usage{
certificate_codes()
}
\value{
list of certificate of completion codes
}
\description{
This function returns a list of current certificate of completion codes from FLDOE
}
\section{Last Updated By}{

Stevens Dormezil
}

\section{Last Update Date}{

2022/06/09
}

\examples{
certificate_codes()

}
\keyword{certificate,}
\keyword{completion}
