% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diploma_codes.R
\name{diploma_codes}
\alias{diploma_codes}
\title{List of FLDOE diploma codes}
\usage{
diploma_codes()
}
\value{
list of diploma codes
}
\description{
This function returns a list of current FLDOE diploma codes with character strings. Contains all diplomas.
}
\section{Last Updated By}{

Stevens Dormezil
}

\section{Last Update Date}{

2022/06/09
}

\examples{
diploma_codes()

}
\keyword{diploma,}
\keyword{graduation}
