% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropout_codes.R
\name{dropout_codes}
\alias{dropout_codes}
\title{List of FLDOE dropout codes.}
\usage{
dropout_codes()
}
\value{
list of dropout codes
}
\description{
This function returns a list of current dropout codes according to federal guidelines.
}
\section{Last Updated By}{

Stevens Dormezil
}

\section{Last Update Date}{

2022/06/09
}

\examples{
dropout_codes()

}
\keyword{dropout}
