% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fldoegrad-package.R
\docType{package}
\name{fldoegrad-package}
\alias{fldoegrad}
\alias{fldoegrad-package}
\title{fldoegrad: FLDOE graduation/diploma helper functions.}
\description{
This package provides a set of functions to help calculate graduation and diploma status according to FLDOE standards.
}
\author{
\strong{Maintainer}: Stevens Dormezil \email{stevens.dormezil@palmbeachschools.org} (\href{https://orcid.org/0000-0003-1886-6964}{ORCID})

}
\keyword{internal}
