% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grad_codes.R
\name{grad_codes}
\alias{grad_codes}
\title{List of federal adult diploma codes.}
\usage{
grad_codes()
}
\value{
list of federal  codes
}
\description{
This function returns a list of current graduation codes according to federal guidelines.
}
\section{Last Updated By}{

Stevens Dormezil
}

\section{Last Update Date}{

2022/06/09
}

\examples{
grad_codes()

}
\keyword{adult,}
\keyword{diploma}
\keyword{federal,}
\keyword{graduation,}
